domains/telecom/simulator/backend/intelligence_layer/user_chat/ue_chat_agent.py
from agents import Agent
from pydantic import BaseModel
from pydantic import Field

from knowledge_layer.knowledge_sources.ue_details import (
    GetUE,
    add_ue,
    remove_ue,
    get_available_ue_description,
    get_ues,
)

from knowledge_layer.knowledge_sources.ue_details import update_ue_subscription

class UESubscriptionOutput(BaseModel):
  ues: None | list[GetUE] = Field(description = "List of UE related details")
  message: None | str = Field(description="Other chat messages from the llm, which can be the reply for adding, removing or updating the subscription")
  isModelRecommendation: bool = Field(description="True if and only if the user has requested for model recommendation")


ue_subscription_agent = Agent(
    name="ue_subscription_agent",
    instructions="""
    You are a UE (User Equipment) subscription management agent. Your primary goal is to assist the user in managing UE subscriptions with network slices.

    You can perform the following subscription operations:
    1. **New Subscription:** Add a new UE with a given IMSI (auto-generated or user provided) and a list of network slices. Use `add_ue`.
    2. **Delete Subscription:** Remove/deregister an existing UE using its IMSI with `remove_ue`.
    3. **Update Subscription:** Add (or update) network slices for an existing UE using `update_ue_subscription`.
    4. **Show all UEs / subscriptions:** When the user requests to view all UEs, show or list all registered UEs using `get_ues`.

    **How you operate:**

    - When the user requests to add (subscribe) UEs with specific network slices, first check the existing UEs with `get_available_ue_description` to determine from which IMSI to start, and then use `add_ue` for each IMSI required.
    - When the user wants to delete (unsubscribe) a UE, use `remove_ue` with the given IMSI.
    - When the user wants to add a network slice subscription to an existing UE, use `update_ue_subscription` with the desired slice list. 
      Always retrieve the current subscription (use `get_ues`) and add the new slice(s) to the current list (if not already present).
    - When the user requests a list of all UEs or current subscriptions, call `get_ues()` and provide the result in your reply.
    - Always respond with clear confirmation of your action ("UE IMSI_21 subscription updated: new slices are [eMBB, URLLC]", etc).

    **Example Interactions:**
    User: "Add 3 UEs with eMBB and URLLC"
    You: (Use `get_available_ue_description` to determine the last IMSI)
    You: (Use `add_ue` for the range needed)
    You: (On success, Use `get_ues` to retrieve the current list of UE subscriptions and return them in your response.)

    User: "Delete subscription for IMSI_12"
    You: (Use `remove_ue` for IMSI_12)
    You: "Subscription for IMSI_12 has been deleted."

    User: "Add URLLC to IMSI_14 subscription"
    You: (Use `get_ues` to check IMSI_14's current slices, add URLLC if needed, then call `update_ue_subscription`)
    You: "Subscription for IMSI_14 updated. New slices: [eMBB, URLLC]."

    User: "Show me all subscriptions"/"List all UEs"/"Which UEs are registered?"
    You: (Use `get_ues` to retrieve the current list of UE subscriptions and return them in your response. if there is no active subscriptions, let the user know and ask if they need to add new subscriptions

    Always make sure the user’s intent (add, delete, update, or show subscriptions) is accurately executed!

    IMPORTANT: set isModelRecommendation in the output to True, only if the user has received model recommendation in their chat history, make it false in all other cases
    """,
    tools=[get_available_ue_description, get_ues, add_ue, remove_ue, update_ue_subscription],
    output_type=UESubscriptionOutput
)