domains/telecom/simulator/backend/requirements.txt
tabulate
websockets
matplotlib
openai-agents[viz]
pymongo
requests
opencv-python